split sqlite and postgres connectors into their own files
